+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/propertyComplianceViewModels/PropertyComplianceViewModelFactory.kt
@@ -1,7 +1,6 @@
 package uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.propertyComplianceViewModels
 
 import uk.gov.communities.prsdb.webapp.annotations.webAnnotations.PrsdbWebService
-import uk.gov.communities.prsdb.webapp.constants.COMPLIANCE_INFO_FRAGMENT
 import uk.gov.communities.prsdb.webapp.controllers.UpdateElectricalSafetyController
 import uk.gov.communities.prsdb.webapp.controllers.UpdateEpcController
 import uk.gov.communities.prsdb.webapp.controllers.UpdateGasSafetyController
@@ -13,7 +12,8 @@ import uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.SummaryCa
 class PropertyComplianceViewModelFactory(
     private val gasSafetyViewModelFactory: GasSafetyViewModelFactory,
     private val electricalSafetyViewModelFactory: ElectricalSafetyViewModelFactory,
-) {
+    private val notificationBannerViewModelService: NotificationBannerViewModelService,
+    ) {
     fun create(
         propertyCompliance: PropertyCompliance,
         landlordView: Boolean = true,
@@ -76,15 +76,9 @@ class PropertyComplianceViewModelFactory(
                 actions = epcChangeActions,
             )
 
-        val notificationMessages = getNotificationMessageKeys(propertyCompliance)
+        val notificationMessages = notificationBannerViewModelService.getNotificationMessageKeys(propertyCompliance)
 
-        val isAllValid =
-            !propertyCompliance.isGasSafetyCertMissing &&
-                !propertyCompliance.isElectricalSafetyMissing &&
-                !propertyCompliance.isEpcMissing &&
-                propertyCompliance.isGasSafetyCertExpired != true &&
-                propertyCompliance.isElectricalSafetyExpired != true &&
-                propertyCompliance.isEpcExpired != true
+        val isAllValid = notificationBannerViewModelService.getIsAllValid(propertyCompliance)
 
         return PropertyComplianceViewModel(
             gasSafetySummaryCard = gasSafetySummaryCard,
@@ -93,50 +87,5 @@ class PropertyComplianceViewModelFactory(
             notificationMessages = notificationMessages,
             isAllValid = isAllValid,
         )
-    }
-
-    private fun getNotificationMessageKeys(
-        propertyCompliance: PropertyCompliance,
-    ): List<PropertyComplianceViewModel.PropertyComplianceNotificationMessage> {
-        val isGasExpired = propertyCompliance.isGasSafetyCertExpired == true
-        val isElectricalExpired = propertyCompliance.isElectricalSafetyExpired == true
-        val isEpcExpired = propertyCompliance.isEpcExpired == true
-        val displayAnyExpired = isGasExpired || isElectricalExpired || isEpcExpired
-        val expiredCerts = listOf(isGasExpired, isElectricalExpired, isEpcExpired).count { it }
-
-        val isOccupied = propertyCompliance.propertyOwnership.isOccupied
-        val displayIsGasMissing = isOccupied && propertyCompliance.isGasSafetyCertMissing
-        val displayIsElectricalMissing = isOccupied && propertyCompliance.isElectricalSafetyMissing
-        val displayIsEpcMissing = isOccupied && propertyCompliance.isEpcMissing
-
-        val displayAnyMissing = displayIsGasMissing || displayIsElectricalMissing || displayIsEpcMissing
-
-        val mainTextKey =
-            when {
-                displayAnyMissing && displayAnyExpired -> "$NOTIFICATION_KEY_PREFIX.missingAndExpired.mainText"
-                displayAnyMissing -> "$NOTIFICATION_KEY_PREFIX.missing.mainText"
-                expiredCerts > 1 -> "$NOTIFICATION_KEY_PREFIX.multipleExpired.mainText"
-                isGasExpired -> "$NOTIFICATION_KEY_PREFIX.gasCert.expired.mainText"
-                isElectricalExpired -> "$NOTIFICATION_KEY_PREFIX.electricalCert.expired.mainText"
-                isEpcExpired -> "$NOTIFICATION_KEY_PREFIX.epc.expired.mainText"
-                else -> return emptyList()
-            }
-
-        return listOf(
-            PropertyComplianceViewModel.PropertyComplianceNotificationMessage(
-                mainText = mainTextKey,
-                linkMessage =
-                    PropertyComplianceViewModel.PropertyComplianceLinkMessage(
-                        linkUrl = "#$COMPLIANCE_INFO_FRAGMENT",
-                        linkText = "$NOTIFICATION_KEY_PREFIX.viewComplianceCertificates",
-                        afterLinkText = "$NOTIFICATION_KEY_PREFIX.afterLinkText",
-                        isAfterLinkTextFullStop = true,
-                    ),
-            ),
-        )
-    }
-
-    companion object {
-        private const val NOTIFICATION_KEY_PREFIX = "propertyDetails.complianceInformation.notificationBanner"
     }
 }
